Architecture: created folders for each ui component
