+++ app/src/main/java/com/breens/mpesaappuiclone/presentation/discover/FragmentDiscover.kt
@@ -1,4 +1,4 @@
-package com.breens.mpesaappuiclone.presentation
+package com.breens.mpesaappuiclone.presentation.discover
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -12,12 +12,8 @@ import com.breens.mpesaappuiclone.R
 private const val ARG_PARAM1 = "param1"
 private const val ARG_PARAM2 = "param2"
 
-/**
- * A simple [Fragment] subclass.
- * Use the [FragmentGrow.newInstance] factory method to
- * create an instance of this fragment.
- */
-class FragmentGrow : Fragment() {
+
+class FragmentDiscover : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -35,7 +31,7 @@ class FragmentGrow : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_grow, container, false)
+        return inflater.inflate(R.layout.fragment_discover, container, false)
     }
 
     companion object {
@@ -45,12 +41,12 @@ class FragmentGrow : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment FragmentGrow.
+         * @return A new instance of fragment FragmentDiscover.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            FragmentGrow().apply {
+            FragmentDiscover().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
