+++ app/src/main/java/com/breens/mpesaappuiclone/presentation/grow/FragmentGrow.kt
@@ -1,4 +1,4 @@
-package com.breens.mpesaappuiclone.presentation
+package com.breens.mpesaappuiclone.presentation.grow
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -12,8 +12,12 @@ import com.breens.mpesaappuiclone.R
 private const val ARG_PARAM1 = "param1"
 private const val ARG_PARAM2 = "param2"
 
-
-class FragmentDiscover : Fragment() {
+/**
+ * A simple [Fragment] subclass.
+ * Use the [FragmentGrow.newInstance] factory method to
+ * create an instance of this fragment.
+ */
+class FragmentGrow : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -31,7 +35,7 @@ class FragmentDiscover : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_discover, container, false)
+        return inflater.inflate(R.layout.fragment_grow, container, false)
     }
 
     companion object {
@@ -41,12 +45,12 @@ class FragmentDiscover : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment FragmentDiscover.
+         * @return A new instance of fragment FragmentGrow.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            FragmentDiscover().apply {
+            FragmentGrow().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
