+++ app/src/main/java/com/breens/mpesaappuiclone/presentation/interface_adapters/ViewPagerAdapter.kt
@@ -1,14 +1,9 @@
-package com.breens.mpesaappuiclone.presentation
+package com.breens.mpesaappuiclone.presentation.interface_adapters
 
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import android.widget.ImageView
-import android.widget.TextView
 import android.widget.Toast
-import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.recyclerview.widget.RecyclerView
-import com.breens.mpesaappuiclone.R
 import com.breens.mpesaappuiclone.data.model.CardInfo
 import com.breens.mpesaappuiclone.databinding.ItemCardviewBinding
