+++ app/src/main/res/layout/fragment_splash.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.FragmentSplash">
+        tools:context=".presentation.splash.FragmentSplash">
 
         <ImageView
             android:id="@+id/logo"
