+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/fragmentSplash"
-        android:name="com.breens.mpesaappuiclone.presentation.FragmentSplash"
+        android:name="com.breens.mpesaappuiclone.presentation.splash.FragmentSplash"
         android:label="fragment_splash"
         tools:layout="@layout/fragment_splash" >
         <action
@@ -22,17 +22,17 @@
         tools:layout="@layout/fragment_home" />
     <fragment
         android:id="@+id/fragmentTransact"
-        android:name="com.breens.mpesaappuiclone.presentation.FragmentTransact"
+        android:name="com.breens.mpesaappuiclone.presentation.transact.FragmentTransact"
         android:label="fragment_transact"
         tools:layout="@layout/fragment_transact" />
     <fragment
         android:id="@+id/fragmentDiscover"
-        android:name="com.breens.mpesaappuiclone.presentation.FragmentDiscover"
+        android:name="com.breens.mpesaappuiclone.presentation.discover.FragmentDiscover"
         android:label="fragment_discover"
         tools:layout="@layout/fragment_discover" />
     <fragment
         android:id="@+id/fragmentGrow"
-        android:name="com.breens.mpesaappuiclone.presentation.FragmentGrow"
+        android:name="com.breens.mpesaappuiclone.presentation.grow.FragmentGrow"
         android:label="fragment_grow"
         tools:layout="@layout/fragment_grow" />
 </navigation>
